// apps/web/src/lib/types.ts
export interface CalendarEvent {
  id: string;
  title: string;
  courseName: string;
  courseCode: string;
  groupCode: string;
  groupId: string;
  groupType: string | null;
  professors: Array<{ id: number; name: string }> | null;
  classroom: string | null;
  campusName: string | null;
  color: string;
  start: Date;
  end: Date;
  courseId: string;
  group: number;
  credits: number;
}